# docs/api/openapi-deno.yaml
openapi: 3.0.3
info:
  title: RITU API (Deno on Cloud Run)
  version: 1.0.0
  description: >
    REST API for the RITU routine tracking app.
    Implemented with Deno on Cloud Run. Backed by Firestore (Native).
servers:
  - url: https://ritu-api-xxxxxxxx.a.run.app/v1
    description: Cloud Run (prod)
  - url: http://localhost:8787/v1
    description: Local (deno run / deno task dev)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    LimitParam:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  schemas:
...
    Routine:
      type: object
      required: [id, userId, title, schedule, visibility, createdAt]
      properties:
        id: { type: string }
        userId: { type: string }
        title: { type: string }
        description: { type: string, nullable: true }
        schedule:
          type: object
          additionalProperties: true
          example: { type: daily }
        autoShare: { type: boolean, default: false }
        visibility: { type: string, enum: [private, public, followers], default: private }
        currentStreak: { type: integer, minimum: 0 }
        maxStreak: { type: integer, minimum: 0 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        deletedAt: { type: string, format: date-time, nullable: true }
    Completion:
      type: object
      required: [id, routineId, userId, date, createdAt]
      properties:
        id: { type: string }
        routineId: { type: string }
        userId: { type: string }
        date: { type: string, format: date }
        createdAt: { type: string, format: date-time }
...
security:
  - bearerAuth: []
paths:
...
  /routines:
    get:
      summary: List my routines
      parameters: [ { $ref: '#/components/parameters/PageParam' }, { $ref: '#/components/parameters/LimitParam' } ]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/Routine' } }
                  page: { type: integer }
                  limit: { type: integer }
                  total: { type: integer }
    post:
      summary: Create routine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string }
                description: { type: string }
                schedule: { type: object, additionalProperties: true }
                autoShare: { type: boolean }
                visibility: { type: string, enum: [private, public, followers] }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Routine' } } } }
  /routines/{id}:
    get:
      summary: Get routine
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Routine' } } } } }
    patch:
      summary: Update routine
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                description: { type: string }
                schedule: { type: object, additionalProperties: true }
                autoShare: { type: boolean }
                visibility: { type: string, enum: [private, public, followers] }
      responses: { '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/Routine' } } } } }
    delete:
      summary: Soft delete routine
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '204': { description: No Content } }
  /routines/{id}/restore:
    post:
      summary: Restore routine within 7 days
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: Restored, content: { application/json: { schema: { $ref: '#/components/schemas/Routine' } } } } }
  /routines/{id}/completions:
    get:
      summary: Get completions in range
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/Completion' } }
    post:
      summary: Record completion
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [date]
              properties:
                date: { type: string, format: date }
      responses: { '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Completion' } } } } }
  /routines/{id}/completions/{date}:
    delete:
      summary: Delete completion by date
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: date, in: path, required: true, schema: { type: string, format: date } }
      responses:
        '204': { description: No Content }
...